+++ bin/multi-memory.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import { main } from "../src/cli/fractal-memory.ts";
+import { main } from "../src/cli/multi-memory.ts";
 
 main().then((code) => {
   process.exitCode = code;